# Constantize messages