--- a/src/ameba/rule/method_names.cr
+++ b/src/ameba/rule/method_names.cr
@@ -43,6 +43,8 @@ module Ameba::Rule
       description = "Enforces method names to be in underscored case"
     end
 
+    MSG = "Method name should be underscore-cased: %s, not %s"
+
     def test(source)
       AST::Visitor.new self, source
     end
@@ -50,8 +52,7 @@ module Ameba::Rule
     def test(source, node : Crystal::Def)
       return if (expected = node.name.underscore) == node.name
 
-      source.error self, node.location,
-        "Method name should be underscore-cased: #{expected}, not #{node.name}"
+      source.error self, node.location, MSG % [expected, node.name]
     end
   end
 end