--- a/src/ameba/rule/predicate_name.cr
+++ b/src/ameba/rule/predicate_name.cr
@@ -45,7 +45,7 @@ module Ameba::Rule
         alternative = $2
         return unless alternative =~ /^[a-z][a-zA-Z_0-9]*$/
 
-        source.error self, node.location, MSG % [alternative, node.name]
+        source.error self, node.location, sprintf(MSG, alternative, node.name)
       end
     end
   end