--- a/src/ameba/rule/type_names.cr
+++ b/src/ameba/rule/type_names.cr
@@ -68,7 +68,7 @@ module Ameba::Rule
       expected = name.camelcase
       return if expected == name
 
-      source.error self, node.location, MSG % [expected, name]
+      source.error self, node.location, sprintf(MSG, expected, name)
     end
 
     def test(source, node : Crystal::ClassDef)