--- a/src/ameba/rule/unless_else.cr
+++ b/src/ameba/rule/unless_else.cr
@@ -48,13 +48,15 @@ module Ameba::Rule
       description = "Disallows the use of an `else` block with the `unless`"
     end
 
+    MSG = "Favour if over unless with else"
+
     def test(source)
       AST::Visitor.new self, source
     end
 
     def test(source, node : Crystal::Unless)
       return if node.else.nop?
-      source.error self, node.location, "Favour if over unless with else"
+      source.error self, node.location, MSG
     end
   end
 end