--- a/src/ameba/rule/variable_names.cr
+++ b/src/ameba/rule/variable_names.cr
@@ -45,7 +45,7 @@ module Ameba::Rule
     private def check_node(source, node)
       return if (expected = node.name.underscore) == node.name
 
-      source.error self, node.location, MSG % [expected, node.name]
+      source.error self, node.location, sprintf(MSG, expected, node.name)
     end
 
     def test(source, node : Crystal::Var)