This should reduce the mem footprint, as a message string is created only once (in a constant) instead of a string per source.