@straight-shoota 2018-04-16
I don't think this should change anything in the compiled binary. Have you actually compared the memory footprint?