@veelenga 2018-04-16
@straight-shoota thanks for looking into it.

Honestly, I didn't do the comparison because it was obvious for me. But you force me to do it, so...

### The main difference

```crystal
require "../src/ameba"
require "benchmark"

source = Ameba::Source.new %()

struct ConstMSGRule < Ameba::Rule::Base
  MSG = "This is the error message"

  def test(source)
    source.error self, 1, 1, MSG
  end
end

struct LiteralMSGRule < Ameba::Rule::Base
  def test(source)
    source.error self, 1, 1, "This is the error message"
  end
end

const_msg = ConstMSGRule.new
liternal_msg = LiteralMSGRule.new

Benchmark.ips do |x|
  x.report("literal") do
    1_000_000.times { liternal_msg.catch(source) }
  end

  x.report("const") do
    1_000_000.times { const_msg.catch(source) }
  end
end

```

produces the following:

```sh
$ crystal bench/constantize_messages.cr --release                                                                                                                         
literal    0.1  (  9.88s ) (± 0.00%)  3.03× slower
  const   0.31  (  3.26s ) (±10.27%)       fastest
```

### Memory footprint

```crystal
require "../src/ameba"
require "benchmark"

source = Ameba::Source.new %()

struct ConstMSGRule < Ameba::Rule::Base
  MSG = "This is the error message"

  def test(source)
    source.error self, 1, 1, MSG
  end
end

struct LiteralMSGRule < Ameba::Rule::Base
  def test(source)
    source.error self, 1, 1, "This is the error message"
  end
end

const_msg = ConstMSGRule.new
liternal_msg = LiteralMSGRule.new

# 1_000_000.times { liternal_msg.catch(source) }
1_000_000.times { const_msg.catch(source) }

```

Measuring via [vmmap](https://developer.apple.com/legacy/library/documentation/Darwin/Reference/ManPages/man1/vmmap.1.html)

**String liternal**

```
Physical footprint:         163.6M
Physical footprint (peak):  172.5M
```

**String constant**
```
Physical footprint:         132.9M
Physical footprint (peak):  132.9M
```

### The binary

Well, in benchmarks you can find the worst case, when the rule always adds the error message. So in reality the difference is not that big.

Running ameba on Crystal repo:

|               | Master branch           | This branch  |
| ------------- |:-------------:| -----:|
| execution time (average) | 1.30 sec | 1.28 sec |
| memory(pick)        | 138.1M      |   134.5M |